✨  feat : Dto record로 변경 (#1)
