+++ src/main/java/org/likelion/zagabi/domain/Value/Controller/ValueController.java
@@ -1,10 +1,10 @@
-package org.likelion.zagabi.Value.Controller;
+package org.likelion.zagabi.domain.Value.Controller;
 
 import lombok.RequiredArgsConstructor;
-import org.likelion.zagabi.Value.Dto.request.CreateValueRequestDto;
-import org.likelion.zagabi.Value.Dto.request.UpdateValueRequestDto;
-import org.likelion.zagabi.Value.Dto.response.ValueResponseDto;
-import org.likelion.zagabi.Value.Service.ValueService;
+import org.likelion.zagabi.domain.Value.Dto.request.CreateValueRequestDto;
+import org.likelion.zagabi.domain.Value.Dto.request.UpdateValueRequestDto;
+import org.likelion.zagabi.domain.Value.Dto.response.ValueResponseDto;
+import org.likelion.zagabi.domain.Value.Service.ValueService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
